[REFACTOR] Convert login form to function declaration and move login form into organisms
