+++ src/components/pages/login.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@/components/atoms/typography.tsx'
-import LoginForm from '@/components/features/login-form.tsx'
+import LoginForm from '@/components/organisms/login-form.tsx'
 import { Footer } from '@/components/molecules/footer.tsx'
 import PageHeader from '@/components/molecules/page-header.tsx'
